refactor: refactor state type import
